% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_custom_domino_legends_categorical}
\alias{create_custom_domino_legends_categorical}
\title{Create custom legends for the domino plot}
\usage{
create_custom_domino_legends_categorical(
  contrast_levels,
  var_positions,
  var_id,
  contrast,
  categorical_colors,
  color_scale_name,
  legend_text_size = 8,
  left_rect_color = "lightblue",
  right_rect_color = "lightpink"
)
}
\arguments{
\item{contrast_levels}{A character vector of contrast level names.}

\item{var_positions}{A data frame containing variable positions.}

\item{var_id}{A string representing the column name for the variable identifier.}

\item{contrast}{A string representing the column name for the contrast variable.}

\item{categorical_colors}{A named vector specifying the colors for each category.}

\item{color_scale_name}{A string specifying the name of the color scale in the legend.}

\item{legend_text_size}{A numeric value indicating the text size for the legend.}

\item{left_rect_color}{A string specifying the color for the left rectangles.}

\item{right_rect_color}{A string specifying the color for the right rectangles.}
}
\value{
A ggplot object containing custom legends.
}
\description{
Create custom legends for the domino plot
}
\keyword{internal}
