% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_custom_legends}
\alias{create_custom_legends}
\title{Create Custom Legends}
\usage{
create_custom_legends(
  data,
  cat_c,
  group,
  cat_c_colors,
  group_colors,
  var_positions,
  num_vars,
  dot_size
)
}
\arguments{
\item{data}{The original data frame.}

\item{cat_c}{The name of the \code{cat_c} variable.}

\item{group}{The name of the group variable.}

\item{cat_c_colors}{A named vector of colors for \code{cat_c}.}

\item{group_colors}{A named vector of colors for the group variable.}

\item{var_positions}{Data frame with variable positions.}

\item{num_vars}{Number of variables in \code{cat_c}.}

\item{dot_size}{The size of the dots used in the plot.}
}
\value{
A combined ggplot object of the custom legends.
}
\description{
Creates custom legend plots for \code{cat_c} and \code{group}.
}
