% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_simple_box_data}
\alias{prepare_simple_box_data}
\title{Prepare Simple Box Data (no grouping)}
\usage{
prepare_simple_box_data(data, cat_a, cat_b, cat_a_order, cat_b_order)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{cat_a}{The name of the column representing category A.}

\item{cat_b}{The name of the column representing category B.}

\item{cat_a_order}{A vector specifying the order of category A.}

\item{cat_b_order}{A vector specifying the order of category B.}
}
\value{
A data frame with box boundaries for plotting.
}
\description{
Prepares data for plotting boxes without grouping by calculating box boundaries based on category positions.
}
