% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diceplot.R
\name{dice_plot}
\alias{dice_plot}
\title{Dice Plot Visualization}
\usage{
dice_plot(
  data,
  cat_a,
  cat_b,
  cat_c,
  group = NULL,
  group_alpha = 0.5,
  title = NULL,
  cat_c_colors = NULL,
  group_colors = NULL,
  custom_theme = theme_minimal(),
  max_dot_size = 5,
  min_dot_size = 2
)
}
\arguments{
\item{data}{A data frame containing the categorical and group variables for plotting.}

\item{cat_a}{A string representing the column name in \code{data} for the first categorical variable.}

\item{cat_b}{A string representing the column name in \code{data} for the second categorical variable.}

\item{cat_c}{A string representing the column name in \code{data} for the third categorical variable.}

\item{group}{A string representing the column name in \code{data} for the grouping variable.}

\item{group_alpha}{A numeric value for the transparency level of the group rectangles. Default is \code{0.5}.}

\item{title}{An optional string for the plot title. Defaults to \code{NULL}.}

\item{cat_c_colors}{A named vector of colors for \code{cat_c} categories. Defaults to \code{NULL} (automatically generated).}

\item{group_colors}{A named vector of colors for the group variable. Defaults to \code{NULL} (automatically generated).}

\item{custom_theme}{A ggplot2 theme for customizing the plot's appearance. Defaults to \code{theme_minimal()}.}

\item{max_dot_size}{Maximal dot size for the plot to scale the dot sizes.}

\item{min_dot_size}{Minimal dot size for the plot to scale the dot sizes.}
}
\value{
A ggplot object representing the dice plot.
}
\description{
This function generates a custom plot based on three categorical variables and a group variable. It adapts to the number of unique categories in \code{cat_c} and allows customization of various plot aesthetics.
}
